Former-commit-id: cbc9dfc10122843968c839412a326945a63aea06

+++ htdocs/bimpcommercial/objects/BimpComm.class.php
@@ -77,6 +77,17 @@ class BimpComm extends BimpDolObject
         return (int) parent::canEditField($field_name);
     }
 
+    public function canEditCommercial()
+    {
+        global $user;
+
+        if ($user->admin || $user->rights->bimpcommercial->edit_commercial) {
+            return 1;
+        }
+
+        return 0;
+    }
+
     // Getters booléens: 
 
     public function isDeletable($force_delete = false)
@@ -592,7 +603,7 @@ class BimpComm extends BimpDolObject
             case 'id_product':
                 $line = $this->getLineInstance();
                 $alias = $line::$parent_comm_type . '_det';
-                if(!$excluded){
+                if (!$excluded) {
                     $joins[$alias] = array(
                         'alias' => $alias,
                         'table' => $line::$dol_line_table,
@@ -605,11 +616,10 @@ class BimpComm extends BimpDolObject
                     $filters[$alias . '.fk_product'] = array(
                         $key => $values
                     );
-                }
-                else{
+                } else {
                     $alias .= '_not';
                     $filters['a.' . $this->getPrimary()] = array(
-                        'not_in' => '(SELECT '.$alias . '.' . $line::$dol_line_parent_field.' FROM '.MAIN_DB_PREFIX.$line::$dol_line_table.' '.$alias . ' WHERE '.$alias . '.fk_product'.' IN ('.implode(',',$values).'))'
+                        'not_in' => '(SELECT ' . $alias . '.' . $line::$dol_line_parent_field . ' FROM ' . MAIN_DB_PREFIX . $line::$dol_line_table . ' ' . $alias . ' WHERE ' . $alias . '.fk_product' . ' IN (' . implode(',', $values) . '))'
                     );
                 }
                 break;
@@ -683,7 +693,7 @@ class BimpComm extends BimpDolObject
                 $alias = 'product_ef';
                 $line = $this->getLineInstance();
                 $line_alias = $line::$parent_comm_type . '_det';
-                if(!$excluded){
+                if (!$excluded) {
                     $joins[$line_alias] = array(
                         'alias' => $line_alias,
                         'table' => $line::$dol_line_table,
@@ -698,11 +708,10 @@ class BimpComm extends BimpDolObject
                     $filters[$alias . '.' . $field_name] = array(
                         ($excluded ? 'not_' : '') . 'in' => $values
                     );
-                }
-                else{
+                } else {
                     $alias .= '_not';
                     $filters['a.' . $this->getPrimary()] = array(
-                        'not_in' => '(SELECT '.$line_alias . '.' . $line::$dol_line_parent_field.' FROM '.MAIN_DB_PREFIX.$line::$dol_line_table.' '.$line_alias . ', '.MAIN_DB_PREFIX.'product_extrafields '.$alias . ' WHERE '.$alias . '.fk_object = ' . $line_alias . '.fk_product AND '.$alias . '.' . $field_name.' IN ('.implode(',',$values).'))'
+                        'not_in' => '(SELECT ' . $line_alias . '.' . $line::$dol_line_parent_field . ' FROM ' . MAIN_DB_PREFIX . $line::$dol_line_table . ' ' . $line_alias . ', ' . MAIN_DB_PREFIX . 'product_extrafields ' . $alias . ' WHERE ' . $alias . '.fk_object = ' . $line_alias . '.fk_product AND ' . $alias . '.' . $field_name . ' IN (' . implode(',', $values) . '))'
                     );
                 }
                 break;
@@ -2403,7 +2412,6 @@ class BimpComm extends BimpDolObject
         return $errors;
     }
 
-
     public function copyRemisesGlobalesFromOrigin($origin, &$errors = array(), $inverse_price = false)
     {
         if ($this->isLoaded($errors)) {
@@ -3158,6 +3166,12 @@ class BimpComm extends BimpDolObject
                         if (!$type_contact && static::$internal_contact_type_required) {
                             $errors[] = 'Type de contact non spécifié';
                         }
+
+                        $id_type_commercial = (int) $this->db->getValue('c_type_contact', 'rowid', 'source = \'internal\' AND element = \'' . $this->dol_object->element . '\' AND code = \'SALESREPFOLL\'');
+                        if ($type_contact == $id_type_commercial && !$this->canEditCommercial()) {
+                            $errors[] = 'Vous n\'avez pas la permission de changer le commercial ' . $this->getLabel('of_a');
+                        }
+
                         if (!count($errors)) {
                             if ($this->dol_object->add_contact($id_user, $type_contact, 'internal') <= 0) {
                                 $errors[] = BimpTools::getMsgFromArray(BimpTools::getErrorsFromDolObject($this->dol_object), 'Echec de l\'ajout du contact');
@@ -3187,8 +3201,14 @@ class BimpComm extends BimpDolObject
             if (!isset($data['id_contact']) || !(int) $data['id_contact']) {
                 $errors[] = 'Contact à supprimer non spécifié';
             } else {
-                if ($this->dol_object->delete_contact((int) $data['id_contact']) <= 0) {
-                    $errors[] = BimpTools::getMsgFromArray(BimpTools::getErrorsFromDolObject($this->dol_object), 'Echec de la suppression du contact');
+                $id_type_contact = (int) $this->db->getValue('element_contact', 'fk_c_type_contact', 'rowid = ' . $data['id_contact']);
+                $id_type_commercial = (int) $this->db->getValue('c_type_contact', 'rowid', 'source = \'internal\' AND element = \'' . $this->dol_object->element . '\' AND code = \'SALESREPFOLL\'');
+                if ($id_type_contact == $id_type_commercial && !$this->canEditCommercial()) {
+                    $errors[] = 'Vous n\'avez pas la permission de changer le commercial ' . $this->getLabel('of_a');
+                } else {
+                    if ($this->dol_object->delete_contact((int) $data['id_contact']) <= 0) {
+                        $errors[] = BimpTools::getMsgFromArray(BimpTools::getErrorsFromDolObject($this->dol_object), 'Echec de la suppression du contact');
+                    }
                 }
             }
         }
